requirements.txt
cachetools==5.5.0
clear==2.0.0
colorama==0.4.6
google-auth==2.35.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
oauthlib==3.2.2
prettytable==3.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
requests-oauthlib==2.0.0
rsa==4.9
